third_party/blink/web_tests/external/wpt/css/css-sizing/keyword-sizes-on-replaced-element-expected.txt

This is a testharness.js-based test.
Found 4 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] .test 68
  assert_equals: \n<canvas width="100" height="100" class="test min-height" style="min-height: fit-content" data-expected-width="10" data-expected-height="10"></canvas>\nwidth expected 10 but got 110
[FAIL] .test 69
  assert_equals: \n<canvas width="100" height="100" class="test max-height" style="max-height: fit-content" data-expected-width="510" data-expected-height="510"></canvas>\nwidth expected 510 but got 110
[FAIL] .test 71
  assert_equals: \n<canvas width="100" height="100" class="test min-height" style="min-width: 50px; min-height: fit-content" data-expected-width="60" data-expected-height="60"></canvas>\nwidth expected 60 but got 110
[FAIL] .test 75
  assert_equals: \n<canvas width="100" height="100" class="test max-height" style="max-width: 150px; max-height: fit-content" data-expected-width="160" data-expected-height="160"></canvas>\nwidth expected 160 but got 110
Harness: the test ran to completion.
